// Swiftpay Front end novo/swiftpay-web-checkout/types/checkout.ts
export interface CheckoutProduct {
  productId: string;
  variantId: string | null;
  name: string;
  description: string | null;
  variantName: string | null;
  type: ProductType;
  imageUrl: string | null;
  imageUrls: string[] | null;
  price: number;
  quantity: number;
  maxQuantity: number | null;
  displayOrder: number;
}